+++ srml/support/procedural/src/storage/genesis_config/genesis_config_def.rs
@@ -1,0 +1,138 @@
+// Copyright 2017-2019 Parity Technologies (UK) Ltd.
+// This file is part of Substrate.
+
+// Substrate is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Substrate is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Substrate.  If not, see <http://www.gnu.org/licenses/>.
+
+use srml_support_procedural_tools::syn_ext as ext;
+use proc_macro2::TokenStream;
+use syn::parse_quote;
+use quote::quote;
+use super::super::{DeclStorageDefExt, StorageLineTypeDef};
+
+pub struct GenesisConfigFieldDef {
+	pub doc: Vec<syn::Meta>,
+	pub name: syn::Ident,
+	pub typ: syn::Type,
+	pub default: TokenStream,
+}
+
+pub struct GenesisConfigDef {
+	pub is_generic: bool,
+	pub fields: Vec<GenesisConfigFieldDef>,
+	pub genesis_struct_decl: TokenStream,
+	pub genesis_struct: TokenStream,
+	pub genesis_impl: TokenStream,
+	pub genesis_where_clause: Option<syn::WhereClause>,
+}
+
+impl GenesisConfigDef {
+	pub fn from_def(def: &DeclStorageDefExt) -> Self {
+		let fields = Self::get_genesis_config_field_defs(def);
+
+		let is_generic = fields.iter()
+			.any(|field| ext::type_contains_ident(&field.typ, &def.module_runtime_generic));
+
+		let (
+			genesis_struct_decl,
+			genesis_impl,
+			genesis_struct,
+			genesis_where_clause
+		) = if is_generic {
+			let runtime_generic = &def.module_runtime_generic;
+			let runtime_trait = &def.module_runtime_trait;
+			let optional_instance = &def.optional_instance;
+			let optional_instance_bound = &def.optional_instance_bound;
+			let optional_instance_bound_optional_default = &def.optional_instance_bound_optional_default;
+			let where_clause = &def.where_clause;
+			(
+				quote!(<#runtime_generic: #runtime_trait, #optional_instance_bound_optional_default>),
+				quote!(<#runtime_generic: #runtime_trait, #optional_instance_bound>),
+				quote!(<#runtime_generic, #optional_instance>),
+				where_clause.clone(),
+			)
+		} else {
+			(quote!(), quote!(), quote!(), None)
+		};
+
+		Self {
+			is_generic,
+			fields,
+			genesis_struct_decl,
+			genesis_struct,
+			genesis_impl,
+			genesis_where_clause,
+		}
+	}
+
+	fn get_genesis_config_field_defs(def: &DeclStorageDefExt) -> Vec<GenesisConfigFieldDef> {
+		let mut config_field_defs = Vec::new();
+
+		for (config_field, line) in def.storage_lines.iter()
+			.filter_map(|line| line.config.as_ref().map(|config_field| (config_field.clone(), line)))
+		{
+			let value_type = &line.value_type;
+
+			let typ = match &line.storage_type {
+				StorageLineTypeDef::Simple(_) => (*value_type).clone(),
+				StorageLineTypeDef::Map(map) | StorageLineTypeDef::LinkedMap(map) => {
+					let key = &map.key;
+					parse_quote!( Vec<(#key, #value_type)> )
+				},
+				StorageLineTypeDef::DoubleMap(map) => {
+					let key1 = &map.key1;
+					let key2 = &map.key2;
+
+					parse_quote!( Vec<(#key1, #key2, #value_type)> )
+				},
+			};
+
+			let default = line.default_value.as_ref()
+				.map(|d| {
+					if line.is_option {
+						quote!( #d.unwrap_or_default() )
+					} else {
+						quote!( #d )
+					}
+				})
+				.unwrap_or_else(|| quote!( Default::default() ));
+
+			config_field_defs.push(GenesisConfigFieldDef {
+				doc: line.doc_attrs.clone(),
+				name: config_field,
+				typ,
+				default,
+			});
+		}
+
+		for line in &def.extra_genesis_config_lines {
+			let doc = line.attrs.iter()
+				.filter_map(|a| a.parse_meta().ok())
+				.filter(|m| m.name() == "doc")
+				.collect();
+
+			let default = line.default.as_ref().map(|e| quote!( #e ))
+				.unwrap_or_else(|| quote!( Default::default() ));
+
+
+			config_field_defs.push(GenesisConfigFieldDef {
+				doc,
+				name: line.name.clone(),
+				typ: line.typ.clone(),
+				default,
+			});
+		}
+
+		config_field_defs
+	}
+}
